Create single subscription with multiple tags
Uses a single callback for all tags with a variant object, so correct
casting must be on the other side with the name of the tag provided in the
arguments of the callback.
Requires a unique name for the subscription on the server.
Returns a single unsubscribe action for all of the monitored items.