Return multiple write results
This returns the error along with the name of the tag, the API before
wouldn't be able to tell which tag write failed.
From testing behavior what I've found is that tag write failures are often
timeout failures that will block the whole return. This can't be solved by
indiviual error messages, but it's better than the inablitly to check for
specific errors.